Rename cicada-core to cicada-mcp-core (PyPI name available)
The name 'cicada-core' is already taken on PyPI by another project.
Rename to 'cicada-mcp-core' to avoid the conflict while preserving
the original monorepo architecture.

Changes:
- Rename packages/cicada-core → packages/cicada-mcp-core
- Rename import cicada_core → cicada_mcp_core
- Update all dependencies and workspace config
- Update publish workflow to publish cicada-mcp-core

This preserves the lightweight dependency structure where cicada-scip
only depends on cicada-mcp-core (not the full cicada-mcp package).
