+++ packages/cicada-mcp-core/src/cicada_mcp_core/formatter_interface.py
