+++ packages/cicada-mcp-core/src/cicada_mcp_core/utils/__init__.py
@@ -1,13 +1,13 @@
 """Utilities for cicada-core."""
 
-from cicada_core.utils.hash_utils import (
+from cicada_mcp_core.utils.hash_utils import (
     compute_file_hash,
     compute_hashes_for_files,
     detect_file_changes,
     load_file_hashes,
     save_file_hashes,
 )
-from cicada_core.utils.storage import (
+from cicada_mcp_core.utils.storage import (
     create_storage_dir,
     get_config_path,
     get_hashes_path,
