+++ packages/cicada-scip/pyproject.toml
@@ -13,7 +13,7 @@ authors = [
 ]
 requires-python = ">=3.10"
 dependencies = [
-    "cicada-core>=0.1.0",
+    "cicada-mcp-core>=0.1.0",
     "protobuf>=4.0.0",
 ]
